pyproject.toml
[project]
name = "novel-maker"
version = "0.1.0"
description = "Multi-agent novel writing system powered by Gemini"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "pydantic>=2.0",
    "python-dotenv",
    "rich",
    "tenacity",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "sse-starlette",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
novel-maker = "novel_maker.main:main"
